move vector_field_function to utils
